+++ apps/web/src/components/editor/table/periods-table.tsx
