+++ apps/web/src/components/editor/table/rooms-table.tsx
