+++ apps/web/src/components/editor/table/subjects-table.tsx
