+++ apps/web/src/components/editor/table/teachers-table.tsx
